+++ Crawler/__init__.py
@@ -1,0 +1,4 @@
+from run_crawler import main
+
+if __name__ == '__main__':
+    main()
